[FA-2] Link to NavLink, organized folders
