+++ src/pages/Ingredients/Ingredients.jsx
@@ -1,4 +1,4 @@
-import MobileHeader from "../MobileHeader/MobileHeader";
+import MobileHeader from "../../App/MobileHeader/MobileHeader";
 
 const Ingredients = () => <MobileHeader> Ingredients </MobileHeader>;
